+++ R/AddPoints.R
@@ -139,7 +139,7 @@
 AddPoints <- function(x, y=NULL, z=NULL, zcol=1, crs=NULL,
                       xlim=NULL, ylim=NULL, zlim=NULL, inches=c(0, 0.2),
                       scaling=c("perceptual", "mathematical", "radius"),
-                      bg="#1F1F1FCB", bg.neg=NULL, fg=NA, lwd=0.25,
+                      bg="#1F1F1FCB", bg.neg=NULL, fg=NA, lwd=0.7,
                       cex=0.7, format=NULL, legend.loc="topright",
                       inset=0.02, bty=c("o", "n"), breaks=NULL, break.labels=NULL,
                       quantile.breaks=FALSE, make.intervals=FALSE,