+++ NEWS.md
@@ -1,5 +1,15 @@
 # inlmisc 0.4.5.9000
 
+- In `Grid2Polygons`, improved performance.
+
+- In `RecreateLibrary`, remove `github` option that allowed package installation from GitHub repositories.
+
+- Add `PrintHelpPages` function, used to print the HTML code associated with
+  help pages of a loaded R package.
+
+- In `BuildVignettes` function, rearrange order of arguments and add `doc` argument,
+  used to specify the directory to write vignette output files.
+
 - In `GetColors` function, add `"bpy"` (blue-pink-yellow) color scheme.
 
 - In `AddPoints` function, change interval to reflect labels, that is,